Fix compilation errors: enum overflow, variable shadowing, and missing method declarations

+++ Source/Eldara/Networking/NetworkTypes.h
@@ -8,7 +8,7 @@
 // ============================================================================
 
 UENUM(BlueprintType)
-enum class EPacketType : uint8
+enum class EPacketType : uint16
 {
 	// Authentication (0-9)
 	LoginRequest = 0,
